add routes register service
